refactor(workspace): extract file tree clipboard helper
