+++ packages/workspace/src/plugins/filesystemPlugin/front/file-tree/FileTreeView.tsx
@@ -59,6 +59,9 @@ import { events, userMeta } from "../../../../front/events"
 import { filesystemEvents } from "../../shared/events"
 import type { PaneProps } from "../../../../shared/types/panel"
 import type { LeftTabParams } from "../../../../shared/plugins/types"
+import { copyToClipboard } from "./clipboard"
+
+export { copyToClipboard } from "./clipboard"
 
 const loadFileTreeComponent = () =>
   import("./FileTree").then((m) => ({ default: m.FileTree }))
@@ -68,42 +71,6 @@ export function preloadFileTreeComponent(): void {
 }
 
 const FileTree = lazy(loadFileTreeComponent)
-
-/**
- * Copy `text` to the clipboard. Falls back to a hidden-textarea + execCommand
- * when `navigator.clipboard` is unavailable (non-secure contexts: plain http
- * served from a non-localhost IP, file://, etc.).
- */
-export async function copyToClipboard(text: string): Promise<void> {
-  if (typeof navigator !== "undefined" && navigator.clipboard?.writeText) {
-    try {
-      await navigator.clipboard.writeText(text)
-      return
-    } catch {
-      // Some browsers reject when not focused — fall through to legacy path.
-    }
-  }
-  if (typeof document === "undefined") {
-    throw new Error("Clipboard not available")
-  }
-  const ta = document.createElement("textarea")
-  ta.value = text
-  ta.setAttribute("readonly", "")
-  ta.style.position = "fixed"
-  ta.style.top = "-9999px"
-  ta.style.left = "-9999px"
-  ta.style.opacity = "0"
-  document.body.appendChild(ta)
-  let ok = false
-  try {
-    ta.focus()
-    ta.select()
-    ok = !!document.execCommand?.("copy")
-  } finally {
-    document.body.removeChild(ta)
-  }
-  if (!ok) throw new Error("Clipboard not available")
-}
 
 interface ContextMenuState {
   node: FileTreeNode
